refactor(rtc_uenclave/data-ocalls::save_sealed_blob): extract module
